Move `preview-table` into `web-common/components`
