+++ web-common/src/components/virtualized-table/core/Row.svelte
