+++ web-common/src/components/virtualized-table/core/StickyHeader.svelte
