+++ web-common/src/components/virtualized-table/sections/PinnedColumns.svelte
