+++ web-common/src/components/virtualized-table/sections/RowHeaders.svelte
