+++ web-common/src/components/virtualized-table/sections/TableCells.svelte
